+++ hugo/content/playground/utils.ts
@@ -1,3 +1,9 @@
+/******************************************************************************
+ * Copyright 2022 TypeFox GmbH
+ * This program and the accompanying materials are made available under the
+ * terms of the MIT License, which is available in the project root.
+ ******************************************************************************/
+
 import { compressToEncodedURIComponent } from "lz-string";
 
 export async function share(grammar: string, content: string): Promise<void> {
