+++ hugo/package.json
@@ -30,10 +30,11 @@
     "lz-string": "^1.4.4",
     "monaco-editor-workers": "0.34.2",
     "monaco-editor-wrapper": "1.3.2",
-    "vscode-languageserver": "8.0.2",
-    "vite": "^3.2.4" 
+    "vite": "^3.2.4",
+    "vscode-languageserver": "8.0.2"
   },
   "dependencies": {
+    "clsx": "^1.2.1",
     "react": "^18.2.0",
     "react-dom": "^18.2.0"
   }
